+++ src/app/(main)/activity/components/reserve-time/index.tsx
@@ -1,16 +1,11 @@
 'use client';
 
-import { useEffect, useRef, useState } from 'react';
-import Calendar from 'react-calendar';
-import { Value } from 'react-calendar/dist/shared/types.js';
+import { FormCalendar } from '@/app/(main)/activity/components/form-calendar';
 import { ReserveTimeProps } from '@/app/(main)/activity/components/reserve-time/reserveTime.types';
 import { ReserveTimeDropdown } from '@/app/(main)/activity/components/reserve-time-dropdown';
-import { IcCalendar, IcMinus, IcPlus } from '@/shared/assets/icons';
+import { IcMinus, IcPlus } from '@/shared/assets/icons';
 import { BUTTON_VARIANTS } from '@/shared/components/buttons/button/button.constants';
-import { Input } from '@/shared/components/input';
 import { cn } from '@/shared/utils/cn';
-import { formatDateKey } from '@/shared/utils/formatDate';
-import 'react-calendar/dist/Calendar.css';
 
 export function ReserveTime({
   value,
@@ -20,39 +15,9 @@ export function ReserveTime({
   isAddAction = false,
   className,
 }: ReserveTimeProps) {
-  const [isCalendarOpen, setIsCalendarOpen] = useState(false);
-  const calendarContainerRef = useRef<HTMLDivElement>(null);
-
   const handleChange = (key: keyof typeof value, val: string) => {
     onChange({ ...value, [key]: val });
   };
-
-  const handleCalendarOpen = () => {
-    setIsCalendarOpen(!isCalendarOpen);
-  };
-
-  const handleCalendarClick = (val: Value) => {
-    if (val instanceof Date) {
-      handleChange('date', formatDateKey(val));
-      setIsCalendarOpen(false);
-    }
-  };
-
-  useEffect(() => {
-    if (!isCalendarOpen) return;
-    const handleClickOutside = (event: PointerEvent) => {
-      if (!calendarContainerRef.current) return;
-      if (!calendarContainerRef.current.contains(event.target as Node)) {
-        setIsCalendarOpen(false);
-      }
-    };
-
-    document.addEventListener('pointerdown', handleClickOutside);
-
-    return () => {
-      document.removeEventListener('pointerdown', handleClickOutside);
-    };
-  }, [isCalendarOpen]);
 
   return (
     <div
@@ -62,33 +27,11 @@ export function ReserveTime({
       )}
     >
       {/* 날짜 */}
-      <div
-        className="relative w-full md:w-auto md:flex-4"
-        ref={calendarContainerRef}
-      >
-        <Input
-          label={hasLabel ? '날짜' : undefined}
-          placeholder="날짜를 선택해 주세요"
-          value={value.date}
-          onClick={handleCalendarOpen}
-          onKeyDown={(e) => e.key === 'Enter' && handleCalendarOpen()}
-          rightIcon={<IcCalendar className="text-black" />}
-          rightIconClassName="pointer-events-none"
-          className="cursor-pointer"
-          required
-          readOnly
-        />
-        {isCalendarOpen && (
-          <div className="bg-primary-50 z-base absolute mt-2">
-            <Calendar
-              locale="ko-KR"
-              calendarType="gregory"
-              className="form-calendar"
-              onChange={handleCalendarClick}
-            />
-          </div>
-        )}
-      </div>
+      <FormCalendar
+        onChange={handleChange}
+        hasLabel={hasLabel}
+        date={value.date}
+      />
       {/* 시간 */}
       <div className="flex flex-1 flex-wrap items-end gap-2.25 md:flex-3 md:flex-nowrap">
         <div className="flex-1 md:w-31">
